requirements.txt
requests>=2.31.0
pyyaml>=6.0
pymongo>=4.6.0
boto3>=1.34.0
numpy>=1.24.0


